requirements.txt

face_recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
cmake>=3.25.0
dlib>=19.24.0